--- src/main/java/web/GenreServlet.java
@@ -1,34 +1,0 @@
-package web;
-
-import service.api.IGenreService;
-import service.factories.GenreServiceSingleton;
-
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.PrintWriter;
-
-@WebServlet(name = "GenreServlet", urlPatterns = "/genres")
-public class GenreServlet extends HttpServlet {
-
-    private final IGenreService service;
-
-    public GenreServlet() {
-        this.service = GenreServiceSingleton.getInstance();
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws IOException {
-
-        req.setCharacterEncoding("UTF-8");
-        resp.setContentType("text/html; charset=UTF-8");
-        PrintWriter writer = resp.getWriter();
-
-        service.getAll()
-                .forEach(genre -> writer.append(genre.getId() + " - " + genre.getGenre())
-                        .append("<br>"));
-    }
-}
